# workflow/openapi.yaml
...
openapi: 3.1.0
info:
  title: GW/AP Debug Platform Workflow API
  version: 0.9.0
servers:
- url: http://localhost:8000/api/v1
...
paths:
...
  /knowledge-routing/import:
    post:
      tags:
      - knowledge-routing
      summary: Import Markdown For Routing
      operationId: importMarkdownForRouting
      parameters:
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_markdown_for_routing_api_v1_knowledge_routing_import_post'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Import Markdown For Routing Api V1 Knowledge Routing Import Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      x-agent-callable: true
      description: ADMIN/EXPERT only. Stage one inactive DRAFT per Markdown file. Host CLI uses bounded masked sections
        and makes zero backend generative calls. Web routing must use a caller-visible Chat selection bound to the
        initiating user and request configuration. No automatic approval or publication.
...
  /workbench/knowledge:
    get:
      tags:
      - workbench
      summary: Knowledge
      operationId: knowledge_api_v1_workbench_knowledge_get
      parameters:
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      - VIEWER
      x-agent-callable: true
...
components:
  schemas:
...
    Body_import_markdown_for_routing_api_v1_knowledge_routing_import_post:
      properties:
        files:
          items:
            type: string
            contentMediaType: application/octet-stream
          type: array
          title: Files
        relative_paths_json:
          type: string
          title: Relative Paths Json
          default: '[]'
        reasoning_owner:
          type: string
          enum:
          - platform_llm
          - host_cli
          title: Reasoning Owner
          default: platform_llm
        model_profile_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Profile Id
        consent_model_egress:
          type: boolean
          title: Consent Model Egress
          default: false
        trust_level:
          type: string
          title: Trust Level
          default: MEDIUM
        confidentiality:
          type: string
          title: Confidentiality
          default: INTERNAL
      type: object
      required:
      - files
      title: Body_import_markdown_for_routing_api_v1_knowledge_routing_import_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
security:
- ApiKey: []